@@ -4,7 +4,7 @@ version: 0.0.1+5
 homepage: https://github.com/luiscarch11/flutter_expandable
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
