@@ -2,39 +2,51 @@ import 'package:flutter/material.dart';
 
 class Expandable extends StatefulWidget {
   const Expandable({
-    Key key,
-    //Controls which expandable belongs to which expander. Must not be null and should be the same as its Expander's controller
-    @required this.controller,
-    //Widgets that will be expanded
-    @required this.child,
-    //Maxheight for the expandable widgets.
-    @required this.height,
-    //MaxWidth for the expandable widgets.
+    Key? key,
+    required this.controller,
+    required this.child,
+    required this.height,
     this.width,
-    //Set to true if you want to expand horizontally. Defaults to false.
     this.animateWidth = false,
-    //Set to true if you want to expand vertically. Defaults to true.
     this.animateHeight = true,
-    //Curve for expansion and folding animations
     this.curve,
-    //Duration for expansion and folding animations
     this.duration,
   })  : assert((animateWidth && width != null) ||
             (!animateWidth && width == null)),
-        assert((animateHeight && height != null) ||
-            (!animateHeight && height == null)),
-        assert(controller != null),
-        assert(child != null),
         assert(animateHeight || animateWidth),
         super(key: key);
+
+  ///Controls which expandable belongs to which expander. Must not be null and should be the same as its Expander's controller
+  ///
   final AnimationController controller;
-  final double height;
-  final double width;
-  final bool animateWidth;
-  final Curve curve;
-  final Duration duration;
-  final bool animateHeight;
+
+  ///Widgets that will be expanded
+  ///
   final Widget child;
+
+  ///Maxheight for the expandable widgets.
+  ///
+  final double height;
+
+  ///MaxWidth for the expandable widgets.
+  ///
+  final double? width;
+
+  ///Set to true if you want to expand horizontally. Defaults to false.
+  ///
+  final bool animateWidth;
+
+  ///Set to true if you want to expand vertically. Defaults to true.
+  ///
+  final bool animateHeight;
+
+  ///Curve for expansion and folding animations
+  ///
+  final Curve? curve;
+
+  ///Duration for expansion and folding animations
+  ///
+  final Duration? duration;
 
   @override
   _ExpandableState createState() => _ExpandableState();
@@ -42,9 +54,9 @@ class Expandable extends StatefulWidget {
 
 class _ExpandableState extends State<Expandable>
     with SingleTickerProviderStateMixin {
-  AnimationController _controller;
-  Animation<double> tween;
-  Animation<double> tweenWidth;
+  late AnimationController _controller;
+  late Animation<double> tween;
+  late Animation<double> tweenWidth;
   @override
   void dispose() {
     _controller.removeListener(() {});
@@ -61,7 +73,7 @@ class _ExpandableState extends State<Expandable>
     );
     tween = Tween<double>(
       begin: 0,
-      end: widget.height ?? 0,
+      end: widget.height,
     ).animate(_controller);
 
     tweenWidth = Tween<double>(
@@ -75,10 +87,11 @@ class _ExpandableState extends State<Expandable>
   Widget build(BuildContext context) {
     return AnimatedContainer(
       constraints: BoxConstraints(
-          maxHeight: widget.animateHeight ? tween.value : widget.height),
+        maxHeight: widget.animateHeight ? tween.value : widget.height,
+      ),
       width: widget.animateWidth ? tweenWidth.value : widget.width,
       curve: widget.curve ?? Curves.linear,
-      duration: widget.duration ?? widget.controller.duration,
+      duration: widget.duration ?? widget.controller.duration!,
       child: SingleChildScrollView(
         physics: NeverScrollableScrollPhysics(),
         child: widget.child,
